Extract aggregation tests to a separate class
This moves the aggregation tests from `AbstractTestQueries` to a
separate class. These tests cover Presto's engine behavior, so there is
no point in running them against different connectors.

Test methods were moved verbatim, with inevitable slight reordering
(attempting to actually introduce more order). There were two exceptions:

- `testGroupByRepeatedField` and `testGroupByRepeatedField2` were
  merged
- `testAggregationWithProjection` and `testAggregationWithProjection2`
  were merged
