+++ presto-cassandra/src/test/java/com/facebook/presto/cassandra/TestCassandraDistributed.java
@@ -45,18 +45,6 @@ public class TestCassandraDistributed
     }
 
     @Override
-    public void testGroupingSetMixedExpressionAndColumn()
-    {
-        // Cassandra does not support DATE
-    }
-
-    @Override
-    public void testGroupingSetMixedExpressionAndOrdinal()
-    {
-        // Cassandra does not support DATE
-    }
-
-    @Override
     public void testRenameTable()
     {
         // Cassandra does not support renaming tables
